requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
sqlalchemy==2.0.42
mysqlclient==2.2.7